Combining commits, changing emails
Repo is being retroactively changed to use the GitHub account's email. On that note, the last two commits are being changed.

First commit text: Empty shortcut is now stored as null, not '', and on_unload event will just cancel itself if it's called before the user is finished loading

Second commit text: So, "" is supposed to be the correct way to 'unset' a command, and upon reverting back to that, the delete row functionality is working correctly again (that was the cause of the mysterious reappearing commands). Need to test on more platforms to see if I need either custom code for it or, more likely, min_version in browser_specific_settings